adding some update

+++ controller/GuestController.js
@@ -14,7 +14,7 @@ const GetSpecificationsForType = async (req, res) =>{
                                 ON svt.spec_value_id = sv.id AND svt.language_id = l.id
                         WHERE sv.spec_id = s.id AND sv.enable = true
                         ORDER BY CASE WHEN sv.absolute_value ~ '\\d+' THEN cast(sv.absolute_value as 
-                            DOUBLE) ELSE null END ASC, sv.absolute_value ASC
+                            DOUBLE PRECISION) ELSE null END ASC, sv.absolute_value ASC
                 )value) AS values
                 
             FROM specifications s
